change: size_t

+++ math/dynamic_binomial.hpp
@@ -10,29 +10,29 @@ namespace Ku {
  */
 template <class T> class DynamicBinomial {
   private:
-    unsigned n;
+    size_t n;
     std::vector<T> fact;
     std::vector<T> ifact;
 
   public:
     DynamicBinomial() : DynamicBinomial{0U} {}
-    explicit DynamicBinomial(const unsigned _n)
+    explicit DynamicBinomial(const size_t _n)
         : n(_n), fact(_n + 1), ifact(_n + 1) {
-        fact[0] = T{1U};
-        for (unsigned i = 0; i < n; i++) {
+        fact[0] = T{1};
+        for (size_t i = 0; i < n; i++) {
             fact[i + 1] = fact[i] * T{i + 1};
         }
 
-        ifact[n] = T{1U} / fact[n];
+        ifact[n] = T{1} / fact[n];
 
-        for (unsigned i = n; i > 0; i--) {
+        for (size_t i = n; i > 0; i--) {
             ifact[i - 1] = ifact[i] * T{i};
         }
     }
 
     T p(const int a, const int b) const {
         if (b < 0 || a < b) {
-            return T{0U};
+            return T{0};
         }
 
         assert(0 <= a);
@@ -44,7 +44,7 @@ template <class T> class DynamicBinomial {
 
     T c(const int a, const int b) const {
         if (b < 0 || a < b) {
-            return T{0U};
+            return T{0};
         }
 
         assert(b <= static_cast<int>(n));
@@ -54,11 +54,11 @@ template <class T> class DynamicBinomial {
 
     T h(const int a, const int b) const {
         if (a == 0 && b == 0) {
-            return T{1U};
+            return T{1};
         }
 
         if (a <= 0 || b < 0) {
-            return T{0U};
+            return T{0};
         }
 
         return c(a + b - 1, b);
